Issue-13 : added processing to detect border contact and issue feedback in all cases
Previously border collision was implemented with gridline nodes and
collision signals. This created a discrepancy between detecting an
initial collision (doable with the area_entered signal) and further
attempts to move the sliders into the map border. Initially this was
solved by adding logic to detect subsequent "collisions" with positional
data, which was effective except it did not permit recieving the
necessary collision node data to play the proper sound (as with
subsequent detections based on collision, the source borderline node was
not sending a signal with itself as an argument). This was fixed with
this commit by detecting all border "collisions" by analyzing positional
data for the sliders and setting flags approrpiately.